feat: 헤더 섹션 분리 (#16)
