+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/saebom/Documents/wishpool/src/app/layout.tsx":"1","/Users/saebom/Documents/wishpool/src/components/layout/Header/Header.tsx":"2","/Users/saebom/Documents/wishpool/src/components/layout/Header/IconSection.tsx":"3","/Users/saebom/Documents/wishpool/src/components/layout/Header/NavSection.tsx":"4"},{"size":798,"mtime":1749476436904,"results":"5","hashOfConfig":"6"},{"size":704,"mtime":1749477674323,"results":"7","hashOfConfig":"6"},{"size":368,"mtime":1749476711025,"results":"8","hashOfConfig":"6"},{"size":898,"mtime":1749477897251,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","suppressedMessages":"12","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"3o0qfb",{"filePath":"13","messages":"14","suppressedMessages":"15","errorCount":1,"fatalErrorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"16","messages":"17","suppressedMessages":"18","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","suppressedMessages":"21","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/saebom/Documents/wishpool/src/app/layout.tsx",[],[],"/Users/saebom/Documents/wishpool/src/components/layout/Header/Header.tsx",["22","23"],[],"/Users/saebom/Documents/wishpool/src/components/layout/Header/IconSection.tsx",[],[],"/Users/saebom/Documents/wishpool/src/components/layout/Header/NavSection.tsx",[],[],{"ruleId":"24","severity":2,"message":"25","line":3,"column":8,"nodeType":null,"messageId":"26","endLine":3,"endColumn":13},{"ruleId":"27","severity":1,"message":"28","line":16,"column":9,"nodeType":"29","endLine":16,"endColumn":76},"@typescript-eslint/no-unused-vars","'Image' is defined but never used.","unusedVar","@next/next/no-img-element","Using `<img>` could result in slower LCP and higher bandwidth. Consider using `<Image />` from `next/image` or a custom image loader to automatically optimize images. This may incur additional usage or cost from your provider. See: https://nextjs.org/docs/messages/no-img-element","JSXOpeningElement"]
