+++ src/components/layout/Header/IconSection.tsx
@@ -1,0 +1,14 @@
+import Icon from '@/components/common/Icon';
+
+const IconSection = ({ onClick }: { onClick: () => void }) => {
+  return (
+    <div className="flex gap-5">
+      <Icon name="wish" title="찜 아이콘" />
+      <button type="button" onClick={onClick}>
+        <Icon name="menu" title="메뉴 아이콘" />
+      </button>
+    </div>
+  );
+};
+
+export default IconSection;
